+++ src/GoingMy.Shared/SharedServices.cs
@@ -4,9 +4,11 @@ public static class SharedServices
 {
     public const string IdentityApi = "identity-api";
     public const string PostApi = "post-api";
+    public const string ChatApi = "chat-api";
     public const string IdentityDb = "identity-db";
     public const string Postgresql = "postgresql";
     public const string MongoDB = "mongodb";
     public const string PostDb = "post-db";
+    public const string ChatDb = "chat-db";
     public const string WebApp = "web-app";
 }